Adding initial workout data was causing and DataIntegrityViolation
Created a test profile for testing and did not load the initial data for
testing, this prevents any dangling references.

Also tried to enable orphan removal for this so that maybe i could
remove a workout and it would also remove the related exercises.

My best guess is that since I was using the same exercise object in
multiple places they were being used for different workout objects as
well, leading to a situation where I delete one of the workouts and it
removes the reference to the Exercise, which is a problem since that
same Exercise is on another Workout.